requirements.txt
absl-py==2.2.0
accelerate==1.2.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
antialiased-cnns==0.3
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.45.0
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cmake==3.31.6
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==3.4.1
decorator==5.2.1
dill==0.3.8
distlib==0.3.9
einops==0.8.1
ema-pytorch==0.7.7
eval-type-backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.18.0
filterpy==1.4.5
fiona==1.10.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gdown==4.4.0
geopandas==0.12.2
googledrivedownloader==1.1.0
gradio==4.44.1
gradio-client==1.3.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
icecream==2.1.4
identify==2.6.9
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib-metadata==8.6.1
importlib-resources==6.5.2
iniconfig==2.1.0
ipython==8.18.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy-loader==0.4
lit==18.1.8
llvmlite==0.43.0
lmdb==1.6.2
loguru==0.7.3
lpips==0.1.4
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
nodeenv==1.9.1
normflows==1.7.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open-clip-torch==2.24.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
polars==1.30.0
pre-commit==4.2.0
prompt-toolkit==3.0.51
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic-core==2.27.2
pydeprecate==0.3.1
pydub==0.25.1
pygments==2.19.1
pyiqa==0.1.10
pyparsing==3.2.1
pyproj==3.6.1
pyre-extensions==0.0.23
pysocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-fid==0.3.0
pytorch-lightning==1.4.2
pytz==2025.1
pywavelets==1.6.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.11.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.1
stringzilla==3.12.5
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.15.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.0.0
torchmetrics==0.6.0
torchsde==0.2.6
torchvision==0.15.1
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.37.2
triton==2.0.0
typer==0.15.2
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
vision-aided-loss==0.1.0
wcwidth==0.2.13
websockets==12.0
werkzeug==3.1.3
xformers==0.0.18
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
imagecodecs
pytorch_fid
rasterio